fix panic 'send on closed channel'
