+++ test/cloudtest/pkg/commands/main.go
@@ -283,7 +283,6 @@ func (ctx *executionContext) performExecution() error {
 	defer statTicker.Stop()
 
 	healthCheckChannel := RunHealthChecks(ctx.cloudTestConfig.HealthCheck)
-	defer close(healthCheckChannel)
 
 	for len(ctx.tasks) > 0 || len(ctx.running) > 0 {
 		// WE take 1 test task from list and do execution.
@@ -1525,35 +1524,4 @@ func initCmd(rootCmd *cloudTestCmd) {
 }
 
 func initConfig() {
-}
-
-// RunHealthChecks - Start goroutines with health check probes
-func RunHealthChecks(checkConfigs []*config.HealthCheckConfig) chan error {
-	errCh := make(chan error)
-	ready := true
-
-	for i := range checkConfigs {
-		go func(c int) {
-			config := checkConfigs[c]
-			for {
-				interval := time.Duration(config.Interval) * time.Second
-				<-time.After(interval)
-
-				timeoutCtx, cancel := context.WithTimeout(context.Background(), interval)
-				defer cancel()
-
-				for _, cmd := range utils.ParseScript(config.Run) {
-					builder := &strings.Builder{}
-					_, err := utils.RunCommand(timeoutCtx, cmd, "", func(s string) {}, bufio.NewWriter(builder), nil, nil, false)
-					if ready && err != nil {
-						ready = false
-						errCh <- errors.Errorf(config.Message)
-						return
-					}
-				}
-			}
-		}(i)
-	}
-
-	return errCh
 }
